lib/chatbot-api/functions/drive-sync/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
boto3>=1.26.0
PyPDF2>=3.0.0  # For PDF operations
requests>=2.31.0  # For downloading files